VCST-672: Use a custom tree node template for shipment operation to d… (#410)
fix: The tree node that represents shipment operation in order details blade will display delivery date or "-" if not set.
fix: The tree node that represents shipment operation in order details blade will not display monetary amount associated with the shipment if it is not set.
fix: These changes will satisfy a common convenience need with shipments - we want to display delivery date (if set) without the need to drill down the shipment sub-blade, because it is a crucial information in customer support. Also, money amount associated with shipments is often not used, so we want to hide it for such orders.