VCST-1752: Performance Degradation in MS SQL Database
fix: Addressed performance degradation in MS SQL Database when using OPENJSON and CONTAINS. The issue was resolved by setting the CompatibilityLevel to 120. For more details, refer to this link: https://www.virtocommerce.org/t/performance-degradation-in-ms-sql-database-queries-after-upgrading-to-ef-core-8/731.

feat: Added connfiguration in appsettings.json  SqlServer:CompatibilityLevel fto set up the compatibility level for SQL Server. The default value is set to 120.

```json
  "SqlServer": {
    "CompatibilityLevel": 120
  }
```